feat: add authentication for bookmark

+++ bookmark/bootmark.go
@@ -5,12 +5,13 @@ import (
 	"fmt"
 	"time"
 
+	"encore.dev/beta/auth"
 	"encore.dev/storage/sqldb"
 	"encore.dev/types/uuid"
 )
 
 // insert inserts a URL into the database.
-func insert(ctx context.Context, id uuid.UUID, url string, owner string, note string) error {
+func insert(ctx context.Context, id uuid.UUID, url string, owner auth.UID, note string) error {
 	createdAt := time.Now().UTC()
 	_, err := sqldb.Exec(ctx, `
 		INSERT INTO bookmark (id, url, owner, note, created_at)
@@ -21,7 +22,7 @@ func insert(ctx context.Context, id uuid.UUID, url string, owner string, note st
 
 type Bookmark struct {
 	ID         uuid.UUID // unique ID
-	OWNER      string    // owner of the bookmark
+	OWNER      auth.UID  // owner of the bookmark
 	URL        string    // url of the bookmark
 	NOTE       string    // optional note
 	CREATED_AT time.Time // date time of creation
@@ -29,25 +30,36 @@ type Bookmark struct {
 
 type BookmarkParams struct {
 	URL         string // the URL to bookmark
-	Owner       string // the owner of the bookmark
 	Description string // optional description of the bookmark
 }
 
 // Bookmark a URL.
 //
-//encore:api public method=POST path=/bookmark
+//encore:api auth method=POST path=/bookmark
 func CreateBookmark(ctx context.Context, p *BookmarkParams) (*Bookmark, error) {
+	// print current user id to console
 	id, err := uuid.NewV4()
 	fmt.Printf("UUIDv4: %s\n", id)
 	if err != nil {
 		return nil, err
 	}
 
-	if err := insert(ctx, id, p.URL, p.Owner, p.Description); err != nil {
+	user := auth.Data()
+	fmt.Print("user=", user)
+
+	// get current user id
+	uid, ok := auth.UserID()
+	if !ok {
+		return nil, fmt.Errorf("no user id")
+	}
+
+	owner := "github-" + uid
+
+	if err := insert(ctx, id, p.URL, owner, p.Description); err != nil {
 		return nil, err
 	}
 
-	return &Bookmark{ID: id, URL: p.URL, OWNER: p.Owner, NOTE: p.Description}, nil
+	return &Bookmark{ID: id, URL: p.URL, OWNER: owner, NOTE: p.Description}, nil
 }
 
 type GetResponse struct {
@@ -80,11 +92,20 @@ func GetBookmarks(ctx context.Context, id string) (*GetResponse, error) {
 }
 
 // Delete deletes a bookmark.
-// encore:api public method=DELETE path=/bookmark/:id
+// encore:api auth method=DELETE path=/bookmark/:id
 func DeleteBookmark(ctx context.Context, id string) error {
+	// get current user id
+	uid, ok := auth.UserID()
+	if !ok {
+		return fmt.Errorf("no user id")
+	}
+
+	owner := "github-" + uid
+
 	_, err := sqldb.Exec(ctx, `
 		DELETE FROM bookmark
 		WHERE id = $1
-	`, id)
+		AND owner = $2
+	`, id, owner)
 	return err
 }
